requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-Limiter==3.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
dependency-injector==4.41.0